prettier: The tragic reformatting associated with 379f6e039aa9e.
Some changes are easier to mentally consume before they've been prettier'd.

While I'll certainly allow Prettier to do what it wants to do, I've confined
it's changes to this commit for easier reviewability.

Note to the reviewer: this commit is exactly the result of running the
`npm run lint-fix` command on a clean Git working tree.

Ref: 379f6e039aa9ed71bcfa7d9bfbfec5dccbfa2786